dedup pzem logic
